Adds Timeout to Extraction Method
Previously the `seven_zip_archive` extract action relied on the default timeout value of the `shell_out` function (default of 600 seconds). On slower machines the extraction of a 7.5Gb iso image can take over 10 minutes. This change adds an attribute `timeout` allowing users to change the timeout value, it still defaults to 600 seconds